requirements.txt
# Django
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<25.0
django-cors-headers>=4.3,<5.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Configuration
python-decouple>=3.8,<4.0

# Production Server
gunicorn>=22.0,<23.0

# Communications (Twilio)
twilio>=9.0,<10.0

# Utilities
python-dateutil>=2.9,<3.0
